refactored project structure to reduce separation of files and group more logically
