+++ app/TodoApp/components/Footer.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import FilterLink from '../../containers/TodoApp/FilterLink'
+import FilterLink from '../containers/FilterLink'
 
 const Footer = () => (
   <p>
