+++ app/TodoApp/components/TodoApp.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Footer from './Footer'
-import AddTodo from '../../containers/TodoApp/AddTodo'
-import VisibleTodoList from '../../containers/TodoApp/VisibleTodoList'
+import AddTodo from '../containers/AddTodo'
+import VisibleTodoList from '../containers/VisibleTodoList'
 
 const TodoApp = () => (
   <div className='todoApp'>
