+++ app/TodoApp/containers/AddTodo.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { addTodo } from '../../actions/TodoActions'
+import { addTodo } from '../actions/TodoActions'
 
 let AddTodo = ({ dispatch }) => {
   let input
