+++ app/TodoApp/containers/FilterLink.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
-import { setVisibilityFilter } from '../../actions/TodoActions'
+import { setVisibilityFilter } from '../actions/TodoActions'
 import ActiveLink from '../../components/ActiveLink'
 
 const mapStateToProps = (state, ownProps) => {
