+++ app/TodoApp/reducers/VisibilityFilter.js
